terraform-visualizer/requirements.txt
python-hcl2>=4.0.0
flask>=3.0.0
